new process working for time taken
