+++ app/views/custom_results/_form.html.erb
@@ -5,7 +5,7 @@
         <div data-nested-form-target="form">
           <%= f.fields_for :options, Option.new, child_index: 1 do |option_subform| %>
             <div class="d-flex mt-4" data-nested-form-target="line" id="1">
-              <%= option_subform.text_field :input, required: true, autofocus: true, class: "form-field", placeholder: 'Enter an option here...', oninvalid: "this.setCustomValidity('Please fill in this option')", oninput: "setCustomValidity('')"%>
+              <%= option_subform.text_field :input, required: true, autofocus: false, class: "form-field", placeholder: 'Enter an option here...', oninvalid: "this.setCustomValidity('Please fill in this option')", oninput: "setCustomValidity('')"%>
               <div class="d-flex ms-3">
                 <button type="button" data-action="click->nested-form#toggle" class="plus-button" data-nested-form-target="sign"></button>
               </div>
