+++ app/views/movie_results/question.html.erb
@@ -1,4 +1,4 @@
-<div class="container w-75" data-controller="time-saved">
+<div class="container w-75" data-controller="time-taken">
   <h1 class="my-3 text-center">Movie Category page - page 3 in Figma</h1>
   <%= render "form" %>
 </div>
