+++ app/views/movie_results/show.html.erb
@@ -2,4 +2,5 @@
 
 <%= @movie_result.movie.title %>
 <p><%= @movie_result.movie.overview %></p>
+<h5>Based on overthinkers everywhere, you saved <%= @movie_result.time_taken  %> on this decision!</h5>
 <%= link_to "Back to roll", movie_questions_path %>
