+++ db/migrate/20230313195903_add_time_taken_column_to_movie_results.rb
@@ -1,5 +1,5 @@
 class AddTimeTakenColumnToMovieResults < ActiveRecord::Migration[7.0]
   def change
-    add_column :movie_results, :time_taken, :integer
+    add_column :movie_results, :time_taken, :string
   end
 end
