+++ db/schema.rb
@@ -29,7 +29,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_03_13_195903) do
     t.bigint "user_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.integer "time_taken"
+    t.string "time_taken"
     t.index ["movie_id"], name: "index_movie_results_on_movie_id"
     t.index ["user_id"], name: "index_movie_results_on_user_id"
   end
